fix: retry get branch API call if commit hasn't changed
Non fast forward errors sometimes occur because GitHub returns a stale commit
when we fetch the target branch. Usually, this is because the last batch succeeded,
but GitHub has not propagated that info to all its APIs yet,
so the stale commit is the one the last successful batch was pushed on top of.
Therefore we cache the last commit we received from GitHub in the process dictionary below,
and then if the next time we call the GitHub branch API we receive the same commit,
we retry the API call with an exponentially increasing delay, until max_delay is exceeded.