fix: better processing of Co-authored-by lines in squash commits (#21)
We change the generation of the commit message for squash commits so that "Co-authored-by" lines included in the PR body are collected and placed directly before the "Co-authored-by" lines created by bors for the other commits in the PR. This should ensure that there is no unnecessary whitespace that confuses GitHub.

cf. https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/co-authored-by.20bors.20issue

I also fixed some buggy code trying to find user emails which was causing bors to commit PRs with the wrong email address.