feat: Zulip notification on failed batch (#24)
The idea is to find all the places in the Batcher code where the `state` of a `Batch` is modified, and then run a new `send_zulip/2` function. (I just searched for uses of `Batch.changeset` and narrowed down to the ones which mentioned `state`.)

This also fixes the Zulip config after #23. 

(I really should add some better testing for the Zulip functionality. At the moment at least the `build_message` functions are being tested, but `BorsNG.Worker.Zulip.send_message` is skipped in testing.)