feat: bump priority of running batches to 100 (#33)
After this change, `bors r+ p=37` will put a PR at the top of list of waiting batches, rather than interrupting the currently running batch. `bors r+ p=101` will still interrupt the currently running batch.

We raise the priority of batches to 100 when they are set to `:running` (batches priority > 100 will not be affected). In particular, setting priority at the batch level means that the constituent PRs will keep their previous priority. In the default case where all PR priorities are 0, this ensures that newly approved PRs can be added to the split-up pieces of a failed batch.